Добавила Home  и Game  страницы. Было прокинуто событие, которое перекидывает с главной страницы на страницу Game. Подключена библиотека classNames, которая позволяет более гибко писать классы, зависящие от переменных. Добавлены новые компоненты MenuHeader, Menu, Navbar. В MenuHeader подключен hook useState, через него организована работа активности и неактивности меню и навбара между собой. Были сложности, не сразу поняла, почему не отображается кнопка гамбургера, потом в группе общались и разобралась.
