+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\index.js":"1","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\App.js":"2","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\HeaderBlock\\index.js":"3","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\Header\\index.js":"4","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\Layout\\index.js":"5","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\Footer\\index.js":"6","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\PokemonCard\\index.js":"7"},{"size":170,"mtime":1611936367361,"results":"8","hashOfConfig":"9"},{"size":4499,"mtime":1612108065076,"results":"10","hashOfConfig":"9"},{"size":537,"mtime":1611768216462,"results":"11","hashOfConfig":"9"},{"size":402,"mtime":1612108065078,"results":"12","hashOfConfig":"9"},{"size":863,"mtime":1612110329788,"results":"13","hashOfConfig":"9"},{"size":320,"mtime":1611936367328,"results":"14","hashOfConfig":"9"},{"size":2091,"mtime":1612108065090,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},"m2u76v",{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\index.js",[],["31","32"],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\App.js",[],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\HeaderBlock\\index.js",[],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\Header\\index.js",[],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\Layout\\index.js",[],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\Footer\\index.js",[],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\PokemonCard\\index.js",[],{"ruleId":"33","replacedBy":"34"},{"ruleId":"35","replacedBy":"36"},"no-native-reassign",["37"],"no-negated-in-lhs",["38"],"no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\index.js":"1","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\App.js":"2","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\HeaderBlock\\index.js":"3","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\Header\\index.js":"4","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\Layout\\index.js":"5","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\Footer\\index.js":"6","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\PokemonCard\\index.js":"7","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\roots\\Home\\index.js":"8","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\roots\\Game\\index.js":"9","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\MenuHeader\\index.js":"10","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\Menu\\index.js":"11","C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\Navbar\\index.js":"12"},{"size":170,"mtime":1611936367361,"results":"13","hashOfConfig":"14"},{"size":490,"mtime":1612200947221,"results":"15","hashOfConfig":"14"},{"size":537,"mtime":1611768216462,"results":"16","hashOfConfig":"14"},{"size":648,"mtime":1612200859545,"results":"17","hashOfConfig":"14"},{"size":891,"mtime":1612197704218,"results":"18","hashOfConfig":"14"},{"size":320,"mtime":1611936367328,"results":"19","hashOfConfig":"14"},{"size":2100,"mtime":1612198026385,"results":"20","hashOfConfig":"14"},{"size":5754,"mtime":1612203258335,"results":"21","hashOfConfig":"14"},{"size":138,"mtime":1612198758962,"results":"22","hashOfConfig":"14"},{"size":468,"mtime":1612206149163,"results":"23","hashOfConfig":"14"},{"size":1113,"mtime":1612206126902,"results":"24","hashOfConfig":"14"},{"size":663,"mtime":1612205984220,"results":"25","hashOfConfig":"14"},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},"m2u76v",{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"28"},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"43","messages":"44","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"45","messages":"46","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"47","messages":"48","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"49","messages":"50","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\index.js",[],["51","52"],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\App.js",[],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\HeaderBlock\\index.js",[],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\Header\\index.js",[],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\Layout\\index.js",[],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\Footer\\index.js",[],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\PokemonCard\\index.js",[],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\roots\\Home\\index.js",[],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\roots\\Game\\index.js",[],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\MenuHeader\\index.js",[],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\Menu\\index.js",[],"C:\\Users\\ageev\\source\\repos\\pokemon-game\\src\\components\\Navbar\\index.js",[],{"ruleId":"53","replacedBy":"54"},{"ruleId":"55","replacedBy":"56"},"no-native-reassign",["57"],"no-negated-in-lhs",["58"],"no-global-assign","no-unsafe-negation"]
