+++ src/components/PokemonCard/index.js
@@ -1,6 +1,9 @@
-import style from './style.module.css';
-import cardBackSide from './images/card-back-side.jpg';
 import { useState } from 'react';
+import cn from 'classnames';
+
+import cardBackSide from './images/card-back-side.jpg';
+
+import style from './style.module.css';
 
 const PokemonCard = ({ name, img, id, type, values }) => {
     const [isActive, setActive] = useState(false);
@@ -9,15 +12,15 @@ const PokemonCard = ({ name, img, id, type, values }) => {
     }
     return (
         <div className={style.root} onClick={handleClick}>
-            <div className={`${style.pokemonCard} ${isActive ? style.active : ''} `}>
+            <div className={cn(style.pokemonCard, { [style.active]: isActive })}>
                 <div className={style.cardFront}>
-                    <div className={`${style.wrap} ${style.front}`}>
-                        <div className={`${style.pokemon} ${style[type]}`}>
+                    <div className={cn(style.wrap, style.front)}>
+                        <div className={cn(style.pokemon, style[type])}>
                             <div className={style.values}>
-                                <div className={`${style.count} ${style.top}`}>{values.top}</div>
-                                <div className={`${style.count} ${style.right}`}>{values.right}</div>
-                                <div className={`${style.count} ${style.bottom}`}>{values.bottom}</div>
-                                <div className={`${style.count} ${style.left}`}>{values.left}</div>
+                                <div className={cn(style.count, style.top)}>{values.top}</div>
+                                <div className={cn(style.count, style.right)}>{values.right}</div>
+                                <div className={cn(style.count, style.bottom)}>{values.bottom}</div>
+                                <div className={cn(style.count, style.left)}>{values.left}</div>
                             </div>
                             <div className={style.imgContainer}>
                                 <img src={img} alt={name} />
@@ -32,7 +35,7 @@ const PokemonCard = ({ name, img, id, type, values }) => {
                 </div>
 
                 <div className={style.cardBack}>
-                    <div className={`${style.wrap} ${style.back}`}>
+                    <div className={cn(style.wrap, style.back)}>
                         <img src={cardBackSide} alt="Сard Backed" />
                     </div>
                 </div>
