+++ src/roots/Game/index.js
@@ -1,0 +1,9 @@
+const GamePage = () => {
+    return (
+        <div>
+            This is GamePage
+        </div>
+    );
+}
+
+export default GamePage;
